+++ src/dom-utils.js
@@ -12,5 +12,12 @@ const renderComponent = (template, container) => {
   container.insertAdjacentHTML(`beforeend`, template);
 };
 
+const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+  return newElement.firstChild;
+};
+
 export {renderContainer};
 export {renderComponent};
+export {createElement};
