+++ src/main.js
@@ -1,7 +1,7 @@
-import {renderContainer} from './components/dom-utils';
-import {renderComponent} from './components/dom-utils';
+import {renderContainer} from './dom-utils';
+import {renderComponent} from './dom-utils';
 
-import {cards} from './data';
+import {tasks} from './data';
 
 import {controls} from './components/controls';
 import {generateConrolsTemplate} from './components/controls';
@@ -12,14 +12,32 @@ import {generateSearchTemplate} from './components/search';
 import {filterNames} from './components/filters';
 import {generateFiltersTemplate} from './components/filters';
 
-import {generateCardsTemplate} from './components/cards';
-import {generateNewCardTemplate} from './components/card-edit';
+import {Task} from './components/task';
+import {TaskEdit} from './components/task-edit';
 
 import {button} from './components/button';
 import {generateButtonTemplate} from './components/button';
 
 const MAX_CARDS_ON_BOARD = 8;
-const [firstCard, ...lastsCards] = cards;
+
+const renderTask = (taskMock, container) => {
+  const task = new Task(taskMock);
+  const taskEdit = new TaskEdit(taskMock);
+
+
+  const taskElement = task.getElement();
+  const taskEditElement = taskEdit.getElement();
+
+  const onTaskElementEdit = () => container.replaceChild(taskEdit.getElement(), task.getElement());
+  const onTaskElementSubmit = () => container.replaceChild(task.getElement(), taskEdit.getElement());
+  const onTaskElementRemove = () => taskEdit.removeElement();
+
+  taskElement.querySelector(`.card__btn--edit`).addEventListener(`click`, onTaskElementEdit);
+  taskEditElement.addEventListener(`submit`, onTaskElementSubmit);
+  taskEditElement.querySelector(`.card__delete`).addEventListener(`click`, onTaskElementRemove);
+
+  task.renderElement(container);
+};
 
 // Main container
 const mainContainer = document.querySelector(`.main`);
@@ -42,13 +60,18 @@ const tasksContainer = renderContainer(`div`, [`board__tasks`], boardContainer);
 renderComponent(generateConrolsTemplate(controls), controlsContainer);
 renderComponent(generateSearchTemplate(search), searchContainer);
 renderComponent(generateFiltersTemplate(filterNames), filtersContainer);
-renderComponent(generateNewCardTemplate(firstCard), tasksContainer);
-renderComponent(generateCardsTemplate(lastsCards.slice(0, MAX_CARDS_ON_BOARD - 1)), tasksContainer);
+
+for (const task of tasks.slice(0, MAX_CARDS_ON_BOARD)) {
+  renderTask(task, tasksContainer);
+}
+
 renderComponent(generateButtonTemplate(button), boardContainer);
 
 const loadMoreButton = document.querySelector(`.load-more`);
 const onLoadMoreButtonClick = () => {
-  renderComponent(generateCardsTemplate(lastsCards.slice(MAX_CARDS_ON_BOARD - 1)), tasksContainer);
+  for (const task of tasks.slice(MAX_CARDS_ON_BOARD)) {
+    renderTask(task, tasksContainer);
+  }
   loadMoreButton.remove();
 };
